backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
httpx==0.27.2
pillow>=11.0.0
numpy>=2.2.3
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
